Remaster/merged detection
- The tool now tries to detect apps that contain a merged patch (remaster or manually merged). Please report any false positives or wrong version numbers.
- New command: [P] - (Hidden) Toggle merged patch detection for app PKGs.
- Added pattern variables %merged_ver% and %true_ver%, which can be used to show a merged patch's version. If a merged patch is found, both %merged_ver% and %true_ver% are the patch version. If no patch is found, %merged_ver% is empty and %true_ver% is %app_ver%.
- The default pattern has been updated to include %merged_ver%.
- Fixed a bug that caused PKGs to be skipped after user input.

+++ pkgrename.c/include/terminal.h
@@ -12,7 +12,7 @@ void reset_terminal();
 void raw_terminal();
 
 // Reads user input and stores it in buffer "string"
-void scan_string(char *string, int max_size, char *default_string,
+void scan_string(char *string, size_t max_size, char *default_string,
   char *(*f)());
 
 #endif
