+++ pkgrename.c/src/releaselists.c
@@ -93,7 +93,7 @@ static int strings_match(char *str1, char *str2) {
     return 0;
   }
 
-  for (int i = 0; i < strlen(str2); i++) {
+  for (size_t i = 0; i < strlen(str2); i++) {
     if (tolower(str2[i]) != tolower(str1[i])) return 0;
   }
